+++ DoAnLTWHQT/Models/ApiLoginRequest.cs
@@ -2,9 +2,8 @@
 {
     public class ApiLoginRequest
     {
-
         public string Email { get; set; }
         public string Password { get; set; }
-
+        public bool RememberMe { get; set; }
     }
 }