+++ DoAnLTWHQT/Models/UpdateCartRequest.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace DoAnLTWHQT.Models
+{
+    public class UpdateCartRequest
+    {
+
+        public int Quantity { get; set; }
+
+    }
+}